requirements.txt
django-htmx==1.0.1
Django==3.2
pillow==9.0.1
requests==2.25.1
